"bugfix": staticcall on non-view functions
apparently the EVM considers that `sload` is a "real" write operation, which makes no sense. Removing "view" modifier and other read-only subroutines fixes the issue.

Hopefully, the only place where we needed to do that were places where we could replace a static call (see `verify` function in ERC2771Forwarder contract) with a real call, because it does not make any actual state change.

Except if you listen to the EVM and consider that transient storage is real storage, that is